requirements.txt
numpy==2.2.6
packaging==26.0
pyserial==3.5
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
shiboken6==6.10.2
